api/requirements.txt
flask
pymongo
joblib
scikit-learn
python-dotenv